Clean up block generator
